+++ packages/runtime/src/mod.ts
@@ -5,6 +5,6 @@
 
 export { mount } from "./mount.ts";
 export { patch } from "./patch/mod.ts";
-export { createDom, replaceNode, setProp, setText } from "./dom/mod.ts";
+export { createDom, removeProp, replaceNode, setProp, setText } from "./dom/mod.ts";
 export { getDomRef, removeDomRef, setDomRef } from "./types.ts";
 export { isEventProp, setEventHandler } from "./events.ts";